imported F.
request.POST is a dictionary type object that lets us acess the data by the key name that is submitted.

request.POST['choice'] will raise a KeyError is the choice was not provided.

F('votes')+1 will increase the count by 1 for the votes in the database.

HttpResponseRedirect is used instead of the HttpResponse because it takes the URL as the argument abd redirects the user. its a good practice.

reverse in the HttpResponseRedirect is used to avoid the hardcoded URL.